Add setting for session timeout.
