--- server/src/helpers/database.ts
@@ -1,13 +1,0 @@
-import { ConnectionOptions } from 'mongoose';
-import config from 'config';
-
-export interface DatabaseConfig {
-  databaseURL: string;
-  secret: string;
-  config: ConnectionOptions;
-  maxRetries?: number;
-}
-
-const databaseConfig: DatabaseConfig = config.get('database');
-
-export default databaseConfig;
